moved the crusade-related entities into a crusade package to keep things clean
