+++ src/main/java/com/crusadehelper/entities/crusade/CrusadeForce.java
@@ -1,4 +1,4 @@
-package com.crusadehelper.entities;
+package com.crusadehelper.entities.crusade;
 
 import com.crusadehelper.enums.faction.Faction;
 import com.fasterxml.jackson.annotation.JsonBackReference;
