+++ src/main/java/com/crusadehelper/services/CrusadeService.java
@@ -1,6 +1,6 @@
 package com.crusadehelper.services;
 
-import com.crusadehelper.entities.Crusade;
+import com.crusadehelper.entities.crusade.Crusade;
 import com.crusadehelper.repositories.crusade.CrusadeRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
